[Separation of Concerns] Molecule
