+++ src/components/Molecule/LinkButtonWithIcon/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" scoped></style>
 <template lang="pug">
   BaseLinkButton(
     :url="url"
@@ -10,42 +12,3 @@
       class="icon"
     )
 </template>
-
-<script>
-import FeatherIcon from '~/components/Atom/FeatherIcon/index.vue'
-import BaseLinkButton from '~/components/Atom/BaseLinkButton/index.vue'
-
-export default {
-  name: 'LinkButtonWithIcon',
-  components: {
-    FeatherIcon,
-    BaseLinkButton,
-  },
-  props: {
-    iconName: {
-      type: String,
-      default: 'TwitterIcon',
-    },
-    iconSize: {
-      type: String,
-      default: '1.5x',
-    },
-    colorType: {
-      type: String,
-      default: 'green',
-    },
-    text: {
-      type: String,
-      require: true,
-      default: 'Text',
-    },
-    url: {
-      type: String,
-      require: true,
-      default: 'https://ja.nuxtjs.org/',
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" scoped />
