+++ src/components/Molecule/LinkTextWithIcon/index.vue
@@ -1,3 +1,5 @@
+<script src="./index.js"></script>
+<style lang="sass" src="./index.sass" scoped></style>
 <template lang="pug">
   BaseLinkText(
     :url="url"
@@ -10,45 +12,3 @@
       class="icon"
     )
 </template>
-
-<script>
-import FeatherIcon from '@/components/Atom/FeatherIcon/index.vue'
-import BaseLinkText from '@/components/Atom/BaseLinkText/index.vue'
-
-export default {
-  name: 'LinkTextWithIcon',
-  components: {
-    FeatherIcon,
-    BaseLinkText,
-  },
-  props: {
-    iconName: {
-      type: String,
-      default: 'TwitterIcon',
-    },
-    iconSize: {
-      type: String,
-      default: '1.5x',
-    },
-    colorType: {
-      type: String,
-      default: 'green',
-      validator(value) {
-        return ['green', 'grey', 'white'].includes(value)
-      },
-    },
-    text: {
-      type: String,
-      require: true,
-      default: 'Text',
-    },
-    url: {
-      type: String,
-      require: true,
-      default: 'https://ja.nuxtjs.org/',
-    },
-  },
-}
-</script>
-
-<style lang="sass" src="./index.sass" scoped />
